4.1.19 move identical functions to task-utils.js
