+++ src/utils/task-utils.js
@@ -1,0 +1,13 @@
+const getDeadlineClass = (dueDate) => {
+  return dueDate instanceof Date && dueDate < Date.now()
+    ? `card--deadline`
+    : ``;
+};
+
+const getRepeatClass = (repeatingDays) => {
+  return Object.values(repeatingDays).some(Boolean)
+    ? `card--repeat`
+    : ``;
+};
+
+export {getDeadlineClass, getRepeatClass};
